+++ src/mimic/orm/macros/src/node/entity_extra.rs
@@ -4,46 +4,38 @@ use crate::{
     node::{Def, MacroNode, Node, Trait, TraitNode, Traits},
 };
 use darling::FromMeta;
-use orm::types::Sorted;
 use orm_schema::Schemable;
 use proc_macro2::TokenStream;
 use quote::quote;
-use syn::{Ident, Path};
+use syn::Path;
 
 ///
-/// Fixture
+/// EntityExtra
 ///
 
 #[derive(Debug, FromMeta)]
-pub struct Fixture {
+pub struct EntityExtra {
     #[darling(default, skip)]
     pub def: Def,
 
     pub entity: Path,
 
-    #[darling(default)]
-    pub sorted: Sorted,
-
-    #[darling(multiple, rename = "key")]
-    pub keys: Vec<Ident>,
+    #[darling(multiple, rename = "source")]
+    pub sources: Vec<EntityExtraSource>,
 }
 
-impl Node for Fixture {
+impl Node for EntityExtra {
     fn expand(&self) -> TokenStream {
-        let Self { sorted, .. } = self;
         let Def { ident, .. } = &self.def;
 
         // quote
         let schema = self.ctor_schema();
         let derive = self.derive();
-        let keys = self.keys.iter().map(quote::ToTokens::to_token_stream);
         let imp = self.imp();
         let q = quote! {
             #schema
             #derive
-            #sorted
             pub enum #ident {
-                #(#keys,)*
             }
             #imp
         };
@@ -58,29 +50,30 @@ impl Node for Fixture {
     }
 }
 
-impl MacroNode for Fixture {
+impl MacroNode for EntityExtra {
     fn def(&self) -> &Def {
         &self.def
     }
 }
 
-impl Schemable for Fixture {
+impl Schemable for EntityExtra {
     fn schema(&self) -> TokenStream {
         let def = &self.def.schema();
         let entity = quote_one(&self.entity, to_path);
-        let keys = quote_vec(&self.keys, to_string);
+        let sources = quote_vec(&self.sources, EntityExtraSource::schema);
 
         quote! {
-            ::mimic::orm::schema::node::SchemaNode::Fixture(::mimic::orm::schema::node::Fixture{
+            ::mimic::orm::schema::node::SchemaNode::EntityExtra(
+                ::mimic::orm::schema::node::EntityExtra{
                 def: #def,
                 entity: #entity,
-                keys: #keys,
+                sources: #sources,
             })
         }
     }
 }
 
-impl TraitNode for Fixture {
+impl TraitNode for EntityExtra {
     fn traits(&self) -> Vec<Trait> {
         let mut traits = Traits::default();
         traits.extend(vec![Trait::Copy, Trait::EnumDisplay, Trait::EnumStaticStr]);
@@ -90,5 +83,43 @@ impl TraitNode for Fixture {
 
     fn map_imp(&self, t: Trait) -> TokenStream {
         imp::any(self, t)
+    }
+}
+
+///
+/// EntityExtraSource
+///
+
+#[derive(Debug, FromMeta)]
+pub struct EntityExtraSource {
+    pub name: String,
+    pub path: Path,
+}
+
+impl Node for EntityExtraSource {
+    fn expand(&self) -> TokenStream {
+        let name = &self.name;
+        let path = &self.path;
+
+        let q = quote! {
+            #name: #path,
+        };
+
+        q
+    }
+}
+
+impl Schemable for EntityExtraSource {
+    fn schema(&self) -> TokenStream {
+        // quote
+        let name = quote_one(&self.name, to_path);
+        let path = quote_one(&self.path, to_string);
+
+        quote! {
+            ::mimic::orm::schema::node::EntityExtraSource {
+                name: #name,
+                path: #path,
+            }
+        }
     }
 }